Move StatusTool to tycho-spi
to make it usable by more plugins.
