+++ tycho-core/src/main/java/org/eclipse/tycho/core/osgitools/BaselineServiceImpl.java
@@ -44,7 +44,7 @@ import org.eclipse.equinox.p2.query.QueryUtil;
 import org.eclipse.equinox.p2.repository.artifact.IArtifactDescriptor;
 import org.eclipse.equinox.p2.repository.artifact.IArtifactRepository;
 import org.eclipse.tycho.MavenRepositoryLocation;
-import org.eclipse.tycho.core.shared.StatusTool;
+import org.eclipse.tycho.helper.StatusTool;
 import org.eclipse.tycho.p2.metadata.IP2Artifact;
 import org.eclipse.tycho.p2.publisher.P2Artifact;
 import org.eclipse.tycho.p2maven.ListQueryable;
