+++ tycho-core/src/main/java/org/eclipse/tycho/p2/resolver/ResolverException.java
@@ -19,7 +19,7 @@ import java.util.stream.Stream;
 
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.equinox.internal.p2.director.Explanation;
-import org.eclipse.tycho.core.shared.StatusTool;
+import org.eclipse.tycho.helper.StatusTool;
 
 /**
  * Exception thrown by Tycho's p2 resolution implementations. It is a checked exception to encourage
